fix: Воспроизведение озвучки в safari
