+++ src/voice-player/index.ts
@@ -1,0 +1,192 @@
+import { createNanoEvents } from '../nanoevents';
+import { VoicePlayerSettings } from '../typings';
+
+import { createTrackStream } from './track-stream';
+import { generateSilence } from './silence-generator';
+
+const AudioContext = window.AudioContext || window.webkitAudioContext;
+
+/** Создает коллекцию треков  */
+const createTrackCollection = <T extends unknown>() => {
+    let trackIds: Array<string>;
+    let trackMap: Map<string, T>;
+
+    const clear = () => {
+        trackIds = new Array<string>();
+        trackMap = new Map<string, T>();
+    };
+
+    const add = (id: string, track: T) => {
+        if (trackMap.has(id)) {
+            throw new Error('Track already exists');
+        }
+
+        trackMap.set(id, track);
+        trackIds.push(id);
+    };
+
+    const has = (id: string) => trackMap.has(id);
+
+    const getById = (id: string): T => {
+        const track = trackMap.get(id);
+        if (track === undefined) {
+            throw new Error('Unknown track id');
+        }
+
+        return track;
+    };
+
+    const getByIndex = (index: number): T => {
+        if (index < 0 || index >= trackIds.length) {
+            throw new Error('Index out of bounds');
+        }
+
+        const track = trackMap.get(trackIds[index]);
+        if (track == null) {
+            throw new Error('Something wrong...');
+        }
+
+        return track;
+    };
+
+    const some = (predicate: (item: T) => boolean) => trackIds.some((id) => predicate(getById(id)));
+
+    clear();
+
+    return {
+        clear,
+        has,
+        get: getById,
+        getByIndex,
+        add,
+        some,
+        get length() {
+            return trackIds.length;
+        },
+    };
+};
+
+export type EventsType = {
+    play: (trackId: string) => void;
+    end: (trackId: string) => void;
+};
+
+export const createVoicePlayer = ({
+    startVoiceDelay = 0.2,
+    sampleRate,
+    numberOfChannels,
+}: VoicePlayerSettings = {}) => {
+    const actx: AudioContext | null = AudioContext ? new AudioContext() : null;
+    const { on, emit } = createNanoEvents<EventsType>();
+    const tracks = createTrackCollection<ReturnType<typeof createTrackStream>>();
+    // true - воспроизводим все треки в очереди (новые в том числе), false - скипаем всю очередь (новые в т.ч.)
+    let active = true;
+    let cursor = 0;
+
+    // если safari - нужно активировать аудиоконтекст по событию ввода
+    if (actx && navigator.vendor.search('Apple') >= 0) {
+        let empty: AudioBuffer | null = null;
+
+        actx.decodeAudioData(generateSilence(), (buffer) => {
+            empty = buffer;
+        });
+
+        const handleClick = () => {
+            document.removeEventListener('click', handleClick);
+            document.removeEventListener('touchstart', handleClick);
+
+            /// нужно что-то проиграть, чтобы сафари разрешил воспроизводить звуки в любой момент в этом контексте
+            /// проигрываем тишину
+            const source = actx.createBufferSource();
+            source.buffer = empty;
+            source.connect(actx.destination);
+            source.start(0);
+        };
+
+        // для пк
+        document.addEventListener('click', handleClick);
+        // для мобильных устройств
+        document.addEventListener('touchstart', handleClick);
+    }
+
+    const play = () => {
+        if (cursor >= tracks.length) {
+            if (tracks.some((track) => !track.loaded)) {
+                return;
+            }
+
+            cursor = 0;
+            tracks.clear();
+            return;
+        }
+
+        const current = tracks.getByIndex(cursor);
+        if (current.status === 'end') {
+            if (cursor < tracks.length) {
+                cursor++;
+                play();
+            }
+        } else {
+            current.play();
+        }
+    };
+
+    const append = (data: Uint8Array, trackId: string, last = false) => {
+        let current = tracks.has(trackId) ? tracks.get(trackId) : undefined;
+        if (current == null) {
+            if (actx == null) {
+                // считаем что аудио неподдерживается, игнорируем вызов
+                return;
+            }
+            current = createTrackStream(actx, {
+                sampleRate,
+                numberOfChannels,
+                delay: startVoiceDelay,
+                onPlay: () => emit('play', trackId),
+                onEnd: () => {
+                    emit('end', trackId);
+                    play();
+                },
+                trackStatus: active ? 'stop' : 'end',
+            });
+            tracks.add(trackId, current);
+        }
+
+        if (data.length) {
+            current.write(data);
+        }
+
+        if (last) {
+            current.setLoaded();
+        }
+
+        play();
+    };
+
+    const stop = () => {
+        while (cursor < tracks.length) {
+            tracks.getByIndex(cursor).stop();
+            cursor++;
+        }
+
+        cursor = 0;
+        tracks.clear();
+    };
+
+    return {
+        append,
+        get active() {
+            return active;
+        },
+        set active(value: boolean) {
+            active = value;
+            if (value) {
+                play();
+            } else {
+                stop();
+            }
+        },
+        on,
+        stop,
+    };
+};
